refactor: move review components to reputation folder and rename GigReviewCard to GigReviewForm
- Move EventReviewSection from components/events to components/reputation
- Move and rename GigReviewCard to GigReviewForm in components/reputation
- Update imports in events/[slug]/page.tsx, gigs/[slug]/page.tsx, and gigs/[slug]/review/page.tsx
- Update component usage from GigReviewCard to GigReviewForm in gig detail and review pages
